+++ custom_components/houseplan/const.py
@@ -34,7 +34,7 @@ PLAN_ORPHAN_TTL_S = 3600
 SCHEDULED_GRACE_S = 30 * 24 * 3600
 FILES_DIR = "houseplan/files"
 CONF_ADMIN_ONLY = "admin_only"
-VERSION = "1.46.5"
+VERSION = "1.46.6"
 
 DEFAULT_CONFIG: dict = {
     "spaces": [],